# make jnp.array faster